Use null-safe operator to accommodate versionless deps
`snapshotDependencyCheck` interrogates dependent modules to see if
their versions are SNAPSHOTs.  Prior to this change, the build would
fail if a dependency did not declare a version.  Now it will tolerate
this.

It is an uncommon use case, but spring-android has it because they
compile against an exploded set of classfiles, which are inherently
without version metadata.