Trigger pom generation on version change
generatePom task now considers gradle.properties as an input
for incremental build purposes.

Prior to this change, one could change the project version
in gradle.properties (as during a release), yet generatePom
would remain UP-TO-DATE, because only changes to the classpath
were considered as inputs to the incremental build.

Now a change to gradle.properties will cause generatePom to
fire, meaning all poms (root and child) will be updated with
the new version.  In this way, generated poms will always
remain in sync with the project version.